% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{beautify}
\alias{beautify}
\title{Make strings prettier for printing}
\usage{
beautify(string)
}
\arguments{
\item{string}{input string}
}
\value{
string
}
\description{
Takes an input string and 'beautify' by converting underscores to spaces and
}
