% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine margin plots with main plot}
\usage{
combine_plots(gg_main, gg_tmb = NULL, gg_gene = NULL)
}
\arguments{
\item{gg_main}{main oncoplot tileplot (ggplot)}

\item{gg_tmb}{barplot describing total mutations. Set to NULL to not draw barplot (ggplot)}

\item{gg_gene}{barplot describing number of mutated samples per gene. Set to NULL to not draw barplot (ggplot)}
}
\value{
patchwork object (or ggplot obj if both \code{gg_tmb} and \code{gg_gene} are NULL)
}
\description{
Combine margin plots with main plot
}
