% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_gene_barplot}
\alias{ggoncoplot_gene_barplot}
\title{Gene barplot}
\usage{
ggoncoplot_gene_barplot(
  .data,
  fontsize_count = 14,
  palette = NULL,
  colour_mutation_type_unspecified = "grey10",
  show_axis
)
}
\arguments{
\item{.data}{data frame output by ggoncoplot_prep_df}

\item{fontsize_count}{fontsize of gene mutation count x axis (number)}

\item{palette}{a named vector mapping all possible mutation types (vector names) to colours (vector values).}

\item{colour_mutation_type_unspecified}{colour of mutations in oncoplot and margin plots if \code{col_mutation_type} is not supplied (string)}

\item{show_axis}{show axis text/ticks/line (flag)}
}
\value{
ggplot showing gene mutation counts
}
\description{
Gene barplot
}
