% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_options}
\alias{ggoncoplot_options}
\title{ggoncoplot options}
\usage{
ggoncoplot_options(
  interactive_svg_width = 12,
  interactive_svg_height = 6,
  plotsize_tmb_rel_height = 10,
  plotsize_gene_rel_width = 20,
  plotsize_metadata_rel_height = 20,
  xlab_title = "Sample",
  ylab_title = "Gene",
  fontsize_xlab = 26,
  fontsize_ylab = 26,
  fontsize_genes = 16,
  fontsize_samples = 12,
  fontsize_count = 14,
  fontsize_tmb_title = 14,
  fontsize_tmb_axis = 11,
  fontsize_pathway = 16,
  tile_height = 1,
  tile_width = 1,
  colour_backround = "grey90",
  colour_mutation_type_unspecified = "grey10",
  show_sample_ids = FALSE,
  show_ylab_title = FALSE,
  show_xlab_title = FALSE,
  show_ylab_title_tmb = FALSE,
  show_axis_gene = TRUE,
  show_axis_tmb = TRUE,
  log10_transform_tmb = TRUE,
  scientific_tmb = FALSE,
  colour_pathway_text = "white",
  colour_pathway_bg = "grey10",
  colour_pathway_outline = "black",
  pathway_text_angle = 0,
  ggoncoplot_guide_ncol = 2
)
}
\arguments{
\item{interactive_svg_width}{dimensions of interactive plot (number)}

\item{interactive_svg_height}{dimensions of interactive plot (number)}

\item{plotsize_tmb_rel_height}{percentage of vertical space TMB margin plot should take up. Must be some value between 5-90 (number)}

\item{plotsize_gene_rel_width}{percentage of horizontal space the gene barplot should take up. Must be some value between 5-90 (number)}

\item{plotsize_metadata_rel_height}{percentage of vertical space the metadata tile plot should take up. Must be some value between 5-90 (number)}

\item{xlab_title}{x axis label. Set \code{xlab_title = NULL} to remove title (string)}

\item{ylab_title}{y axis of interactive plot. Set \code{ylab_title = NULL} to remove title (string)}

\item{fontsize_xlab}{size of x axis title (number)}

\item{fontsize_ylab}{size of y axis title (number)}

\item{fontsize_genes}{size of y axis text (gene names) (number)}

\item{fontsize_samples}{size of x axis text (sample names). Ignored unless show_sample_ids is set to true (number)}

\item{fontsize_count}{fontsize of gene mutation count x axis (number)}

\item{fontsize_tmb_title}{fontsize of y axis title for TMB marginal plot (number)}

\item{fontsize_tmb_axis}{fontsize of y axis text for TMB marginal plot (number)}

\item{fontsize_pathway}{fontsize of y axis strip text describing gene pathways (number)}

\item{tile_height}{proportion of available vertical space each tile will take up (0-1) (number)}

\item{tile_width}{proportion of available horizontal space each tile take up (0-1) (number)}

\item{colour_backround}{colour used for background non-mutated tiles (string)}

\item{colour_mutation_type_unspecified}{colour of mutations in oncoplot and margin plots if \code{col_mutation_type} is not supplied (string)}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}

\item{show_ylab_title}{show y axis title of oncoplot (flag)}

\item{show_xlab_title}{show x axis title of oncoplot (flag)}

\item{show_ylab_title_tmb}{show y axis title of TMB margin plot (flag)}

\item{show_axis_gene}{show x axis line/ticks/labels for gene barplot (flag)}

\item{show_axis_tmb}{show y axis line/ticks/labels for TMB barplot (flag)}

\item{log10_transform_tmb}{log10 transform total number of mutations for TMB marginal plot (flag)}

\item{scientific_tmb}{display tmb counts in scientific notation (flag)}

\item{colour_pathway_text}{colour of text describing pathways (string)}

\item{colour_pathway_bg}{background fill colour of pathway strips (string)}

\item{colour_pathway_outline}{outline colour of pathway strips (string)}

\item{pathway_text_angle}{angle of pathway text label (typically 0 or 90 degrees) (number)}

\item{ggoncoplot_guide_ncol}{how many columns to use when describing oncoplot legend (number)}
}
\value{
ggoncoplot options object ready to be passed to \code{\link[=ggoncoplot]{ggoncoplot()}} \code{options} argument
}
\description{
Customise the look of your oncoplot.
}
\examples{

# Read GBM MAF file
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)
gbm_df <- read.csv(file = gbm_csv, header = TRUE)

# Plot Oncoplot and Customise Options
gbm_df |>
  ggoncoplot(
    col_genes = 'Hugo_Symbol',
    col_samples = 'Tumor_Sample_Barcode',
    col_mutation_type = 'Variant_Classification',

    # Customise Visual Options
    options = ggoncoplot_options(

      # Interactive Plot Options
      interactive_svg_width = 12,
      interactive_svg_height = 6,

      # Relative height of different plotsizes
      plotsize_tmb_rel_height = 10,
      plotsize_gene_rel_width = 20,
      plotsize_metadata_rel_height = 20,

      # Axis Titles
      xlab_title = "Glioblastoma Samples",
      ylab_title = "Top 10 mutated genes",

      # Fontsizes
      fontsize_xlab = 40,
      fontsize_ylab = 40,
      fontsize_genes = 16,
      fontsize_samples = 12,
      fontsize_count = 14,
      fontsize_tmb_title = 14,
      fontsize_tmb_axis = 11,
      fontsize_pathway = 16,

      # Customise Tiles
      tile_height = 1,
      tile_width = 1,
      colour_backround = "grey90",
      colour_mutation_type_unspecified = "grey10",

      # Show different elements
      show_sample_ids = FALSE,
      show_ylab_title = FALSE,
      show_xlab_title = FALSE,
      show_ylab_title_tmb = FALSE,
      show_axis_gene = TRUE,
      show_axis_tmb = TRUE,

      # Transformation and label scales
      log10_transform_tmb = TRUE,
      scientific_tmb = FALSE,

      # Pathway Faceting Colours / Text
      colour_pathway_text = "white",
      colour_pathway_bg = "grey10",
      colour_pathway_outline = "black",
      pathway_text_angle = 0,

      # Legend number of columns
      ggoncoplot_guide_ncol = 2
    )
  )
}
