% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_prep_df}
\alias{ggoncoplot_prep_df}
\title{Prep data for oncoplot}
\usage{
ggoncoplot_prep_df(
  .data,
  col_genes,
  col_samples,
  genes_for_oncoplot,
  col_mutation_type = NULL,
  col_tooltip = col_samples,
  pathway = NULL,
  verbose = TRUE
)
}
\arguments{
\item{.data}{data for oncoplot. A data.frame with 1 row per mutation in your cohort. Must contain columns describing gene_symbols and sample_identifiers (data.frame)}

\item{col_genes}{name of \strong{data} column containing gene names/symbols (string)}

\item{col_samples}{name of \strong{data} column containing sample identifiers (string)}

\item{genes_for_oncoplot}{a list of genes to include in the oncoplot (character).}

\item{col_mutation_type}{name of \strong{data} column describing mutation types (string)}

\item{col_tooltip}{name of \strong{data} column containing whatever information you want to display in (string)}

\item{pathway}{a two column dataframe describing pathway. The column containing gene names should have the same name as \strong{col_gene} (data.frame, optional)}

\item{verbose}{verbose mode (flag, default TRUE)}
}
\value{
dataframe with the following columns: 'Gene', 'Sample', 'MutationType', 'Tooltip'.
Sample is a factor with levels sorted in appropriate order for oncoplot vis.
Genes represents either topn genes or specific genes set by \code{genes_to_include}
}
\description{
Prep data for oncoplot
}
\examples{
#' # ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)

# Get genes in appropriate order for oncoplot
genes_for_oncoplot <- ggoncoplot:::get_genes_for_oncoplot(
  .data = gbm_df,
  col_samples = "Tumor_Sample_Barcode",
  col_genes = "Hugo_Symbol",
  topn = 20,
  verbose = FALSE
)

# Create dataframe basis of oncoplot (1 row per sample-gene combo)
ggoncoplot:::ggoncoplot_prep_df(
  gbm_df,
  col_genes = "Hugo_Symbol",
  col_samples = "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification",
  genes_for_oncoplot = genes_for_oncoplot
)

}
