% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{identify_topn_genes}
\alias{identify_topn_genes}
\title{Identify top genes from a mutation df}
\usage{
identify_topn_genes(
  .data,
  col_samples,
  col_genes,
  topn,
  genes_to_ignore = NULL,
  return_extra_genes_if_tied = FALSE
)
}
\arguments{
\item{.data}{data for oncoplot. A data.frame with 1 row per mutation in your cohort. Must contain columns describing gene_symbols and sample_identifiers (data.frame)}

\item{col_samples}{name of \strong{data} column containing sample identifiers (string)}

\item{col_genes}{name of \strong{data} column containing gene names/symbols (string)}

\item{topn}{how many of the top genes to visualise. Ignored if \code{genes_to_include} is supplied (number)}

\item{genes_to_ignore}{genes that should be ignored (character)}

\item{return_extra_genes_if_tied}{instead of strictly returning \code{topn} genes,
in the case of ties (where multiple genes are mutated in the exact same number of samples, complicating selection of top n genes), return all tied genes (potentially more than topn).
If FALSE, will return strictly \code{topn} genes, breaking ties based on order of appearance in dataset (flag)}
}
\value{
vector of topn genes. Their order will be their rank (most mutated = first) (character)
}
\description{
Identify top genes from a mutation df
}
