% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncosim}
\alias{oncosim}
\title{Simulated Cancer Genome Dataset}
\format{
\subsection{\code{oncosim}}{

A data frame with 143 rows and 3 columns:
\describe{
\item{Samples}{Sample containing mutation in the specified gene}
\item{Genes}{Mutated  gene}
\item{VariantType}{Type of mutation in gene}
}
}
}
\source{
not applicable, simulated
}
\usage{
oncosim
}
\description{
An artificial cancer dataset describing mutations found in 9 different tumour samples.
Rows represent mutations.
}
\keyword{datasets}
