% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncosim_metadata}
\alias{oncosim_metadata}
\title{Simulated Cancer Dataset Metadata}
\format{
\subsection{\code{oncosim_metadata}}{

A data frame with 11 rows and 6 columns:
\describe{
\item{Samples}{Unique sample identifiers}
\item{numeric_feature}{Numeric variable including zeros, positive and negative values, \code{NA}, and \code{Inf}/\code{-Inf}}
\item{categorical_feature4levels}{Categorical variable with four levels (\code{"cat"}, \code{"dog"}, \code{"magpie"}, \code{"giraffe"}), may contain empty strings or \code{NA}}
\item{clinical_feature2levels}{Clinical categorical variable indicating biological sex with two levels (\code{"male"}, \code{"female"}), may contain \code{NA}}
\item{logical_feature}{Logical variable with \code{TRUE}, \code{FALSE}, or \code{NA}}
\item{numeric_that_could_be_logical}{Integer variable coded as \code{0} or \code{1} (and \code{NA}) that could be interpreted as logical}
}
}
}
\source{
not applicable, simulated
}
\usage{
oncosim_metadata
}
\description{
A sample‐level metadata table for the \code{oncosim} simulated cancer dataset.
Contains assorted numeric, categorical, clinical, and logical features for each sample.
}
\keyword{datasets}
