% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{prettify}
\alias{prettify}
\title{Make strings prettier for printing}
\usage{
prettify(string, space_after_apostrophe = TRUE, autodetect_units = TRUE)
}
\arguments{
\item{string}{input string}

\item{space_after_apostrophe}{add a space after any apostrophe so long as its
after an alphanumeric character and followed by anything but a space (flag)}

\item{autodetect_units}{automatically detect units (e.g. mm, kg, etc) and wrap in brackets.}
}
\value{
string
}
\description{
Takes an input string and 'prettify' by converting underscores to spaces, capitalising each word, etc.
}
