% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{rank_genes_based_on_pathways}
\alias{rank_genes_based_on_pathways}
\title{Calculate Pathway-informed Genes Rankings}
\usage{
rank_genes_based_on_pathways(
  gene_pathway_map,
  generanks = unique(as.character(gene_pathway_map[[1]])),
  pathwayranks = unique(as.character(gene_pathway_map[[2]]))
)
}
\arguments{
\item{gene_pathway_map}{dataframe where column 1 = gene names and column 2 = pathway names}

\item{generanks}{gene names in the order they should be ranked, where earlier in vector = further up in oncoplot. (character)}

\item{pathwayranks}{pathway names in the order they should be ranked, where earlier in vector = further up in oncoplot (character)}
}
\value{
gene names, sorted based on order they should appear in oncoplot (first = top). Only returns genes present in generanks (character)
}
\description{
Which genes should appear at the top of the oncoplot?
This function takes pathway and gene ranks and returns a list of genes sorted first by pathway then by gene rank.
Gene & pathway rankings can be calculated upstream. By default will use their order in gene_pathway_map.
}
