% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{unify_samples}
\alias{unify_samples}
\title{Prepare dataset for plotting}
\usage{
unify_samples(data, col_samples, samples_to_show)
}
\arguments{
\item{data}{dataframe with a column describing sample IDs (data.frame)}

\item{col_samples}{name of column in \code{data} containing sample IDs (character)}

\item{samples_to_show}{the samples we want to show in plots.
These samples should be the only ones represented in data.frame content,
and any missing ones will be added as factor levels (character)}
}
\value{
data.frame
}
\description{
Take a dataframe containing a column describing sample IDs (\code{col_sample})
Filter on \code{col_sample} \%in\% samples_to_show.
Add any missing samples_to_show not present DF as levels of \code{col_sample}.
This way, when plotting we can use scale_x_discrete(drop=FALSE) to display all the samples we care about
}
