% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot}
\alias{ggoncoplot}
\title{ggoncoplot}
\usage{
ggoncoplot(
  data,
  col_genes,
  col_samples,
  col_mutation_type = NULL,
  genes_to_include = NULL,
  genes_to_ignore = NULL,
  col_tooltip = col_samples,
  topn = 10,
  return_extra_genes_if_tied = FALSE,
  draw_gene_barplot = FALSE,
  draw_tmb_barplot = FALSE,
  copy = c("sample", "gene", "tooltip", "mutation_type", "nothing"),
  palette = NULL,
  metadata = NULL,
  metadata_palette = NULL,
  col_samples_metadata = col_samples,
  cols_to_plot_metadata = NULL,
  metadata_require_mutations = TRUE,
  pathway = NULL,
  col_genes_pathway = col_genes,
  show_all_samples = FALSE,
  total_samples = c("any_mutations", "all", "oncoplot"),
  interactive = TRUE,
  options = ggoncoplot_options(),
  verbose = TRUE
)
}
\arguments{
\item{data}{data for oncoplot. A data.frame with 1 row per mutation in your cohort. Must contain columns describing gene_symbols and sample_identifiers (data.frame)}

\item{col_genes}{name of \strong{data} column containing gene names/symbols (string)}

\item{col_samples}{name of \strong{data} column containing sample identifiers (string)}

\item{col_mutation_type}{name of \strong{data} column describing mutation types (string, optional)}

\item{genes_to_include}{specific genes to include in the oncoplot (character, optional)}

\item{genes_to_ignore}{names of the genes that should be ignored (character, optional)}

\item{col_tooltip}{name of \strong{data} column containing whatever information you want to display in (string, defaults to col_samples)}

\item{topn}{how many of the top genes to visualize. Ignored if \code{genes_to_include} is supplied (number, default 10)}

\item{return_extra_genes_if_tied}{instead of strictly returning \code{topn} genes,
in the case of ties (where multiple genes are mutated in the exact same number of samples, complicating selection of top n genes), return all tied genes (potentially more than topn).
If FALSE, will return strictly \code{topn} genes, breaking ties based on order of appearance in dataset (flag, default FALSE)}

\item{draw_gene_barplot}{add a barplot describing number of samples with each gene mutated (right side) (flag, default FALSE)}

\item{draw_tmb_barplot}{add a barplot describing total number of mutations in each sample (above main plot). If a single gene is mutated multiple times, all mutations are counted towards total (flag, default FALSE)}

\item{copy}{value to copy to clipboard when an oncoplot tile is clicked (string, one of 'sample', 'gene', 'tooltip', 'mutation_type', 'nothing', default 'sample')}

\item{palette}{a named vector mapping all possible mutation types (vector names) to colors (vector values, optional)}

\item{metadata}{dataframe describing sample level metadata.
One column must contain unique sample identifiers. Other columns can describe numeric / categorical metadata (data.frame, optional)}

\item{metadata_palette}{A list of named vectors. List names correspond to metadata column names (categorical only). Vector names to levels of columns. Vector values are colors, the vector names are used to map values in data to a color. (optional)}

\item{col_samples_metadata}{which column in metadata data.frame describes sample identifiers (string, defaults to col_samples)}

\item{cols_to_plot_metadata}{names of columns in metadata that should be plotted (character, optional)}

\item{metadata_require_mutations}{filter out samples from metadata lacking any mutations in data (flag, default TRUE)}

\item{pathway}{a two column dataframe describing pathway. The column containing gene names should have the same name as \strong{col_gene} (data.frame, optional)}

\item{col_genes_pathway}{which column in pathway data.frame describes gene names (string, defaults to col_genes)}

\item{show_all_samples}{show all samples in oncoplot, even if they don't have mutations in the selected genes. Samples only described in metadata but with no mutations at all are still filtered out by default, but you can show these too by setting \code{metadata_require_mutations = FALSE} (flag, default FALSE)}

\item{total_samples}{Strategy for calculating the total number of samples.
This value is used to compute the proportion of mutation recurrence displayed in the tooltip when hovering over the gene barplot,
or as a text annotation when \code{ggoncoplot_options(show_genebar_labels = TRUE)} is set to TRUE.

Possible values:
\itemize{
\item \strong{any_mutations}: All the samples that are in \code{data} (the mutation dataset), irrespective of whether they are on the oncoplot or not.
\item \strong{oncoplot}: Only the samples that are present on the oncoplot.
\item \strong{all}: All the samples in either \code{data} or \code{metadata}.
}}

\item{interactive}{should plot be interactive (boolean, default TRUE)}

\item{options}{a list of additional visual parameters created by calling \code{\link[=ggoncoplot_options]{ggoncoplot_options()}}. See \code{\link{ggoncoplot_options}} for details.}

\item{verbose}{verbose mode (flag, default TRUE)}
}
\value{
ggplot or girafe object if \code{interactive=TRUE}
}
\description{
ggoncoplot
}
\examples{
# ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_clinical_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_clinical.csv"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)
gbm_clinical_df <- read.csv(file = gbm_clinical_csv, header = TRUE)

# Plot Basic Oncoplot
ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification",
  metadata = gbm_clinical_df,
  cols_to_plot_metadata = "gender"
)

# Customise how the Oncoplot looks
ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification",
  metadata = gbm_clinical_df,
  cols_to_plot_metadata = "gender",

  # Customise Visual Options
  options = ggoncoplot_options(
      xlab_title = "Glioblastoma Samples",
      ylab_title = "Top 10 mutated genes"
  )
)
}
