% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot}
\alias{ggoncoplot}
\title{GG oncoplot}
\usage{
ggoncoplot(
  .data,
  col_genes,
  col_samples,
  col_mutation_type = NULL,
  col_tooltip = col_samples,
  topn = 10,
  show_sample_ids = FALSE,
  interactive = TRUE,
  interactive_svg_width = 10,
  interactive_svg_height = 6,
  genes_to_include = NULL,
  xlab_title = "Sample",
  ylab_title = "Gene",
  sample_annotation_df = NULL
)
}
\arguments{
\item{.data}{data for oncoplot. A data.frame with 1 row per mutation in your cohort. Must contain columns describing gene_symbols and sample_identifiers, (data.frame)}

\item{col_genes}{name of \strong{data_main} column containing gene names/symbols (string)}

\item{col_samples}{name of \strong{data_main} column containing sample identifiers (string)}

\item{col_mutation_type}{name of \strong{data_main} column describing mutation types (string)}

\item{col_tooltip}{name of \strong{data_main} column containing whatever information you want to display in (string)}

\item{topn}{how many of the top genes to visualise. Ignored if \code{genes_to_include} is supplied (number)}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}

\item{interactive}{should plot be interactive (boolean)}

\item{interactive_svg_width}{dimensions of interactive plot (number)}

\item{interactive_svg_height}{dimensions of interactive plot (number)}

\item{genes_to_include}{specific genes to include in the oncoplot (character)}

\item{xlab_title}{x axis lable (string)}

\item{ylab_title}{y axis of interactive plot (number)}

\item{sample_annotation_df}{a data.frame with 1 row per sample, with columns representing metadata to annotate oncoplot with. 1st column must contain sample identifiers (data.frame)}
}
\value{
ggplot or ggiraph object if \code{interactive=TRUE}
}
\description{
GG oncoplot
}
\examples{
# ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)

ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification"
)

}
