% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot}
\alias{ggoncoplot}
\title{GG oncoplot}
\usage{
ggoncoplot(
  .data,
  col_genes,
  col_samples,
  col_mutation_type = NULL,
  genes_to_include = NULL,
  col_tooltip = col_samples,
  topn = 10,
  palette = NULL,
  show_sample_ids = FALSE,
  interactive = TRUE,
  interactive_svg_width = 12,
  interactive_svg_height = 6,
  xlab_title = "Sample",
  ylab_title = "Gene",
  fontsize_xlab = 26,
  fontsize_ylab = 26,
  fontsize_genes = 16,
  fontsize_samples = 12
)
}
\arguments{
\item{.data}{data for oncoplot. A data.frame with 1 row per mutation in your cohort. Must contain columns describing gene_symbols and sample_identifiers, (data.frame)}

\item{col_genes}{name of \strong{data} column containing gene names/symbols (string)}

\item{col_samples}{name of \strong{data} column containing sample identifiers (string)}

\item{col_mutation_type}{name of \strong{data} column describing mutation types (string)}

\item{genes_to_include}{specific genes to include in the oncoplot (character)}

\item{col_tooltip}{name of \strong{data} column containing whatever information you want to display in (string)}

\item{topn}{how many of the top genes to visualise. Ignored if \code{genes_to_include} is supplied (number)}

\item{palette}{a named vector mapping all possible mutation types (vector names) to colours (vector values).
If not supplied ggoncoplot will check if all values are either valid SO or MAF variant classification terms
and use pre-made colour schemes for each of these ontologies from the \strong{mutationtypes} package.
If mutation type terms are not described using these ontologies, a 12 colour RColourBrewer palette will be used, but the user warned to make a custom mapping to force consistent colour schemes between plots (character)}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}

\item{interactive}{should plot be interactive (boolean)}

\item{interactive_svg_width}{dimensions of interactive plot (number)}

\item{interactive_svg_height}{dimensions of interactive plot (number)}

\item{xlab_title}{x axis lable (string)}

\item{ylab_title}{y axis of interactive plot (number)}

\item{fontsize_xlab}{size of x axis title (number)}

\item{fontsize_ylab}{size of y axis title (number)}

\item{fontsize_genes}{size of y axis text (gene names) (number)}

\item{fontsize_samples}{size of x axis text (sample names). Ignored unless show_sample_ids is set to true (number)}
}
\value{
ggplot or girafe object if \code{interactive=TRUE}
}
\description{
GG oncoplot
}
\examples{
# ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)

ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification"
)

}
