% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_gene_barplot}
\alias{ggoncoplot_gene_barplot}
\title{Gene barplot}
\usage{
ggoncoplot_gene_barplot(.data, fontsize_count = 14, palette = NULL)
}
\arguments{
\item{.data}{data frame output by ggoncoplot_prep_df}

\item{fontsize_count}{fontsize of gene mutation count x axis (number)}

\item{palette}{a named vector mapping all possible mutation types (vector names) to colours (vector values).
If not supplied ggoncoplot will check if all values are either valid SO or MAF variant classification terms
and use pre-made colour schemes for each of these ontologies from the \strong{mutationtypes} package.
If mutation type terms are not described using these ontologies, a 12 colour RColourBrewer palette will be used, but the user warned to make a custom mapping to force consistent colour schemes between plots (character)}
}
\value{
ggplot showing gene mutation counts
}
\description{
Gene barplot
}
