% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_plot}
\alias{ggoncoplot_plot}
\title{Plot oncoplot}
\usage{
ggoncoplot_plot(
  .data,
  show_sample_ids = FALSE,
  interactive = TRUE,
  interactive_svg_width = 12,
  interactive_svg_height = 6,
  xlab_title = "Sample",
  ylab_title = "Gene",
  sample_annotation_df = NULL
)
}
\arguments{
\item{.data}{data.frame returned from [ggoncoplot_prep_df()]}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}

\item{interactive}{should plot be interactive (boolean)}

\item{interactive_svg_width}{dimensions of interactive plot (number)}

\item{interactive_svg_height}{dimensions of interactive plot (number)}

\item{xlab_title}{x axis lable (string)}

\item{ylab_title}{y axis of interactive plot (number)}
}
\value{
ggplot or girafe object if \code{interactive=TRUE}
}
\description{
This function takes the output from \strong{ggoncoplot_prep_df} and plots it.
Should not be exposed since it makes some assumptions
}
\examples{
# ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)

ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification"
)

}
