% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_plot}
\alias{ggoncoplot_plot}
\title{Plot oncoplot}
\usage{
ggoncoplot_plot(
  .data,
  show_sample_ids = FALSE,
  palette = NULL,
  xlab_title = "Sample",
  ylab_title = "Gene",
  fontsize_xlab = 16,
  fontsize_ylab = 16,
  fontsize_genes = 14,
  fontsize_samples = 10,
  tile_height = 1,
  tile_width = 1,
  colour_backround = "grey90"
)
}
\arguments{
\item{.data}{transformed data from [ggoncoplot_prep_df()] (data.frame)}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}

\item{palette}{a named vector mapping all possible mutation types (vector names) to colours (vector values).
If not supplied ggoncoplot will check if all values are either valid SO or MAF variant classification terms
and use pre-made colour schemes for each of these ontologies from the \strong{mutationtypes} package.
If mutation type terms are not described using these ontologies, a 12 colour RColourBrewer palette will be used, but the user warned to make a custom mapping to force consistent colour schemes between plots (character)}

\item{xlab_title}{x axis lable (string)}

\item{ylab_title}{y axis of interactive plot (number)}

\item{fontsize_xlab}{size of x axis title (number)}

\item{fontsize_ylab}{size of y axis title (number)}

\item{fontsize_genes}{size of y axis text (gene names) (number)}

\item{fontsize_samples}{size of x axis text (sample names). Ignored unless show_sample_ids is set to true (number)}

\item{tile_height}{proportion of available vertical space each tile will take up (0-1) (number)}

\item{tile_width}{proportion of available horizontal space  each tile take up (0-1) (number)}

\item{colour_backround}{colour used for background non-mutated tiles (string)}
}
\value{
ggplot or girafe object if \code{interactive=TRUE}
}
\description{
This function takes the output from \strong{ggoncoplot_prep_df} and plots it.
Should not be exposed since it makes some assumptions about structure of input data.
}
\examples{
# ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)

ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification"
)

}
