% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{ggoncoplot_plot}
\alias{ggoncoplot_plot}
\title{Plot oncoplot}
\usage{
ggoncoplot_plot(
  data,
  show_sample_ids = FALSE,
  palette = NULL,
  show_ylab_title = FALSE,
  show_xlab_title = FALSE,
  xlab_title = "Sample",
  ylab_title = "Gene",
  sample_id_position = c("bottom", "top"),
  sample_id_angle = 90,
  fontsize_xlab = 16,
  fontsize_ylab = 16,
  fontsize_genes = 14,
  fontsize_samples = 10,
  fontsize_legend_title = 12,
  fontsize_legend_text = 12,
  tile_height = 1,
  tile_width = 1,
  copy = c("sample", "gene", "tooltip", "mutation_type", "nothing"),
  colour_backround = "grey90",
  colour_mutation_type_unspecified = "grey10",
  fontsize_pathway = 16,
  colour_pathway_text = "white",
  colour_pathway_bg = "grey10",
  colour_pathway_outline = "black",
  pathway_text_angle = 0,
  legend_title = "Mutation Type",
  show_legend_titles = TRUE,
  ggoncoplot_guide_ncol = 2,
  legend_key_size = 0.3,
  margin_t = 0.2,
  margin_r = 0.3,
  margin_b = 0.2,
  margin_l = 0.3,
  margin_unit = "cm",
  mutation_type_supplied = TRUE
)
}
\arguments{
\item{data}{transformed data from \code{\link[=ggoncoplot_prep_df]{ggoncoplot_prep_df()}} (data.frame)}

\item{show_sample_ids}{show sample_ids_on_x_axis (flag)}

\item{palette}{a named vector mapping all possible mutation types (vector names) to colors (vector values, optional)}

\item{show_ylab_title}{show y axis title of oncoplot (flag)}

\item{show_xlab_title}{show x axis title of oncoplot (flag)}

\item{xlab_title}{x axis label. Set \code{xlab_title = NULL} to remove title (string)}

\item{ylab_title}{y axis of interactive plot. Set \code{ylab_title = NULL} to remove title (string)}

\item{sample_id_position}{should sample names on the x axis be on the \strong{top} or \strong{bottom} of the main oncoplot (string)}

\item{sample_id_angle}{angle of the sample names (number)}

\item{fontsize_xlab}{size of x axis title (number)}

\item{fontsize_ylab}{size of y axis title (number)}

\item{fontsize_genes}{size of y axis text (gene names) (number)}

\item{fontsize_samples}{size of x axis text (sample names). Ignored unless show_sample_ids is set to true (number)}

\item{fontsize_legend_title}{fontsize of the legend titles  (number)}

\item{fontsize_legend_text}{fontsize of the legend text (number)}

\item{tile_height}{proportion of available vertical space each tile will take up (0-1) (number)}

\item{tile_width}{proportion of available horizontal space each tile take up (0-1) (number)}

\item{copy}{value to copy to clipboard when an oncoplot tile is clicked (string, one of 'sample', 'gene', 'tooltip', 'mutation_type', 'nothing', default 'sample')}

\item{colour_backround}{colour used for background non-mutated tiles (string)}

\item{colour_mutation_type_unspecified}{colour of mutations in oncoplot and margin plots if \code{col_mutation_type} is not supplied (string)}

\item{fontsize_pathway}{fontsize of y axis strip text describing gene pathways (number)}

\item{colour_pathway_text}{colour of text describing pathways (string)}

\item{colour_pathway_bg}{background fill colour of pathway strips (string)}

\item{colour_pathway_outline}{outline colour of pathway strips (string)}

\item{pathway_text_angle}{angle of pathway text label (typically 0 or 90 degrees) (number)}

\item{legend_title}{name of legend title (string)}

\item{show_legend_titles}{show legend titles (flag)}

\item{ggoncoplot_guide_ncol}{how many columns to use when describing oncoplot legend (number)}

\item{legend_key_size}{width of the legend key block (number)}

\item{margin_t, margin_r, margin_b, margin_l}{margin for top, right, bottom, and left side of plot. By default, unit is 'cm' but can be changed by setting \code{margin_unit} to any value \code{\link[ggplot2:element]{ggplot2::margin()}} will understand (number)}

\item{margin_unit}{Unit of margin specification. By default is 'cm' but can be changed by setting \code{margin_unit} to any value \code{\link[ggplot2:element]{ggplot2::margin()}} will understand (string)}

\item{mutation_type_supplied}{did user supply a mutation_type column? If not, will turn off legend.}
}
\value{
ggplot or girafe object if \code{interactive=TRUE}
}
\description{
This function takes the output from \strong{ggoncoplot_prep_df} and plots it.
Should not be exposed since it makes some assumptions about structure of input data.
}
\examples{
# ===== GBM =====
gbm_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_maf.csv.gz"
)

gbm_clinical_csv <- system.file(
  package = "ggoncoplot",
  "testdata/GBM_tcgamutations_mc3_clinical.csv"
)

gbm_df <- read.csv(file = gbm_csv, header = TRUE)
gbm_clinical_df <- read.csv(file = gbm_clinical_csv, header = TRUE)

# Plot Basic Oncoplot
ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification",
  metadata = gbm_clinical_df,
  cols_to_plot_metadata = "gender"
)

# Customise how the Oncoplot looks
ggoncoplot(
  gbm_df,
  "Hugo_Symbol",
  "Tumor_Sample_Barcode",
  col_mutation_type = "Variant_Classification",
  metadata = gbm_clinical_df,
  cols_to_plot_metadata = "gender",

  # Customise Visual Options
  options = ggoncoplot_options(
      xlab_title = "Glioblastoma Samples",
      ylab_title = "Top 10 mutated genes"
  )
)
}
