% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoplot.R
\name{theme_oncoplot_default}
\alias{theme_oncoplot_default}
\title{Oncoplot Theme: default}
\usage{
theme_oncoplot_default(...)
}
\arguments{
\item{...}{passed to [ggplot2::theme()] theme}
}
\description{
Oncoplot Theme: default
}
